requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
decorator==5.1.1
dnspython==2.6.1
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
ipython==8.22.2
jedi==0.19.1
kombu==5.3.5
magic-filter==1.0.12
matplotlib-inline==0.1.6
motor==3.3.2
multidict==6.0.5
parso==0.8.3
pexpect==4.9.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
pymongo==4.6.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.3
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.36.3
traitlets==5.14.2
typing_extensions==4.10.0
tzdata==2024.1
uvicorn==0.28.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
